+++ MatrixSDK/Crypto/Algorithms/Megolm/MXMegolmDecryption.m
@@ -297,6 +297,7 @@
     operation = [crypto ensureOlmSessionsForDevices:@{
                                           userId: @[deviceInfo]
                                           }
+                                              force:NO
                                 success:^(MXUsersDevicesMap<MXOlmSessionResult *> *results)
      {
          MXStrongifyAndReturnIfNil(self);
@@ -323,7 +324,7 @@
 
          NSLog(@"[MXMegolmDecryption] shareKeysWithDevice: sharing keys for session %@|%@ with device %@:%@", senderKey, sessionId, userId, deviceId);
 
-         NSDictionary *payload = [self buildKeyForwardingMessage:roomId senderKey:senderKey sessionId:sessionId];
+         NSDictionary *payload = [self->crypto buildMegolmKeyForwardingMessage:roomId senderKey:senderKey sessionId:sessionId chainIndex:nil];
 
          MXDeviceInfo *deviceInfo = olmSessionResult.device;
 
@@ -461,29 +462,6 @@
     {
         NSLog(@"[MXMegolmDecryption] requestKeysForEvent: ERROR: missing fields in event %@", event);
     }
-}
-
-- (NSDictionary*)buildKeyForwardingMessage:(NSString*)roomId senderKey:(NSString*)senderKey sessionId:(NSString*)sessionId
-{
-    NSDictionary *key = [olmDevice getInboundGroupSessionKey:roomId senderKey:senderKey sessionId:sessionId];
-    if (key)
-    {
-        return @{
-                 @"type": kMXEventTypeStringRoomForwardedKey,
-                 @"content": @{
-                         @"algorithm": kMXCryptoMegolmAlgorithm,
-                         @"room_id": roomId,
-                         @"sender_key": senderKey,
-                         @"sender_claimed_ed25519_key": key[@"sender_claimed_ed25519_key"],
-                         @"session_id": sessionId,
-                         @"session_key": key[@"key"],
-                         @"chain_index": key[@"chain_index"],
-                         @"forwarding_curve25519_key_chain": key[@"forwarding_curve25519_key_chain"]
-                         }
-                 };
-    }
-
-    return nil;
 }
 
 @end